lasted update

+++ delete_student.php
@@ -1,6 +1,12 @@
 <?php require_once 'utils.php';
-if (isset($_POST['id'])) {
-	$id = $_POST['id'];
-	$sql = 'delete from student where id = '.$id;
-	execute($sql);
+session_start();
+if (empty($_SESSION['id']) && empty($_SESSION['username'])) {
+  header("location: index.php");
+}
+else {
+	if (isset($_POST['id'])) {
+		$id = $_POST['id'];
+		$sql = 'delete from student where id = '.$id;
+		execute($sql);
+	}
 }
